A little further consolidation
